Move functions to debug the parser to a separate module.
